feat(react-ui): rename tooltip folder
